services/agents/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
schedule
psycopg2-binary
sqlalchemy
pytz